+++ lib/src/image_downloader.dart
@@ -1,0 +1,19 @@
+import 'dart:io';
+
+import 'package:http/http.dart' as http;
+import 'package:path_provider/path_provider.dart';
+
+class ImageDownloaderService {
+  static Future<String?> downloadImage({
+    required String url,
+    required String fileName,
+  }) async {
+    final directory = await getApplicationDocumentsDirectory();
+    final filePath = '${directory.path}/$fileName';
+    final response = await http.get(Uri.parse(url));
+    final file = File(filePath);
+
+    await file.writeAsBytes(response.bodyBytes);
+    return filePath;
+  }
+}